# docs/swagger.yaml
basePath: /api
definitions:
...
  models.Review:
    properties:
      _id:
        type: string
      description:
        type: string
      productId:
        type: string
      rating:
        type: number
      title:
        type: string
    type: object
...
  models.User:
    properties:
      _id:
        type: string
      email:
        type: string
      f_name:
        type: string
      gender:
        type: string
      l_name:
        type: string
      password:
        type: string
      phone_no:
        type: string
      role:
        type: string
    type: object
host: 13.53.115.94
info:
  contact:
    email: alijabbar0034@gmail.com
    name: Ali Jabbar
    url: https://twitter.com/sntshk
  description: An Ecommerce service API in Go using Gin framework.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gin Ecommerce Web
  version: "1.0"
paths:
...
  /review:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all reviews.
      produces:
      - application/json
      responses:
        "200":
          description: '{"total": integer, "reviews": []}'
          schema:
            type: object
        "500":
          description: '{"message": "Internal Server Error"}'
          schema:
            type: object
      summary: Get All Reviews
      tags:
      - Review
  /review/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a review by ID.
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"count": integer}'
          schema:
            type: object
        "400":
          description: '{"message": "Please Provide ID"}'
          schema:
            type: object
        "500":
          description: '{"message": "Internal Server Error"}'
          schema:
            type: object
      summary: Delete Review
      tags:
      - Review
    get:
      consumes:
      - application/json
      description: Retrieves a review by ID.
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Got successfully", "review": object}'
          schema:
            type: object
        "500":
          description: '{"message": "Internal Server Error"}'
          schema:
            type: object
      summary: Get Review By ID
      tags:
      - Review
    post:
      consumes:
      - application/json
      description: Creates a new review for a product by ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Review object
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.Review'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Review added successfully", "id": string}'
          schema:
            type: object
        "400":
          description: '{"message": "Invalid data"}'
          schema:
            type: object
        "500":
          description: '{"message": "Error during creating review"}'
          schema:
            type: object
      summary: Create Review
      tags:
      - Review
    put:
      consumes:
      - application/json
      description: Updates a review by ID.
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      - description: Review object
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.Review'
      produces:
      - application/json
      responses:
        "200":
          description: '{"updatedCount": integer}'
          schema:
            type: object
        "400":
          description: '{"message": "Provide the valid data"}'
          schema:
            type: object
        "500":
          description: '{"message": "Internal Server Error"}'
          schema:
            type: object
      summary: Update Review
      tags:
      - Review
  /user:
    get:
      description: Retrieves a list of all users.
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "All user list", "total": integer, "users": []}'
          schema:
            type: object
        "404":
          description: '{"message": "No users Found"}'
          schema:
            type: object
      summary: Get All Users
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Updates user information.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Updated successfully", "count": integer}'
          schema:
            type: object
        "400":
          description: '{"message": "Invalid info"}'
          schema:
            type: object
        "500":
          description: '{"message": "Error during updating user"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update User
      tags:
      - User
  /user/{id}:
    get:
      description: Retrieves user information by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "User information"}'
          schema:
            type: object
        "404":
          description: '{"message": "No user Found"}'
          schema:
            type: object
      summary: Get User By Id
      tags:
      - User
...
  /user/profile:
    get:
      consumes:
      - application/json
      description: Retrieves the profile information of the authenticated user.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"user": "User profile information"}'
          schema:
            type: object
        "401":
          description: '{"message": "Unauthorized"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get User Profile
      tags:
      - User
  /user/register:
    post:
      description: Responds with the message and token books as JSON.
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Login Successfully", "token": string}'
          schema:
            type: object
        "400":
          description: '{"message": "Inavalid data"} "Bad request'
          schema:
            type: object
        "401":
          description: '{"message": "No user Found"} "Unauthorized er Error'
          schema:
            type: object
      summary: Login User
      tags:
      - User
swagger: "2.0"